feat: 유저 테스트 결과 model 생성

+++ src/main/java/com/dnd5/timoapi/domain/test/domain/model/UserTestResult.java
@@ -1,8 +1,5 @@
 package com.dnd5.timoapi.domain.test.domain.model;
 
-import static com.dnd5.timoapi.global.security.context.SecurityUtil.getCurrentUserId;
-
-import com.dnd5.timoapi.domain.test.domain.model.enums.TestRecordStatus;
 import com.dnd5.timoapi.domain.test.domain.model.enums.ZtpiCategory;
 import java.time.LocalDateTime;
 
@@ -10,14 +7,14 @@ public record UserTestResult(
         Long id,
         Long testRecordId,
         ZtpiCategory category,
-        float score,
+        double score,
         LocalDateTime createdAt,
         LocalDateTime updatedAt
 ) {
     public static UserTestResult create(
             Long testRecordId,
             ZtpiCategory category,
-            float score
+            double score
     ) {
         return new UserTestResult(
                 null,
